feat: remove tool selection dialog
