+++ src/pages/game/information-blocks/GameInformationBlock.jsx
@@ -1,13 +1,26 @@
+import { useState } from 'react'
 import { useSelector } from 'react-redux'
+import styled from '@emotion/styled'
+import PropTypes from 'prop-types'
+import { Accordion, AccordionSummary, AccordionDetails } from '@mui/material'
+import { Done } from '@mui/icons-material'
 
-import { Chip, HoverTooltip } from '../../../components'
+import { useContextData } from '../../../DataContext'
+import { color } from '../../../styles'
+import { Chip, HoverTooltip, IconButton } from '../../../components'
 import { selectSelectedTool, selectRound } from '../game.slice'
 import { PayoffChart } from '../PayoffChart'
 import { StyledCard, StyledRow } from './styles'
 
-export const GameInformationBlock = () => {
+export const GameInformationBlock = ({ isReadyGetNextRoundTool, onSelectNextRoundTool }) => {
+  const contextData = useContextData()
+  const {
+    data: { toolsAvailable = {} },
+  } = contextData
+
   const selectedTool = useSelector(selectSelectedTool)
   const round = useSelector(selectRound)
+  const [expandedTool, setExpandedTool] = useState(null)
 
   return (
     <StyledCard>
@@ -15,13 +28,37 @@ export const GameInformationBlock = () => {
         <Chip label="回合" />
         <div>{round}</div>
       </StyledRow>
-      <StyledRow>
-        <Chip label="本回合輔助指標" />
-        <div>{selectedTool[selectedTool.length - 1]?.displayName ?? ''}</div>
-        {selectedTool[selectedTool.length - 1]?.introduction && (
-          <HoverTooltip tooltip={selectedTool[selectedTool.length - 1]?.introduction} />
-        )}
-      </StyledRow>
+      {isReadyGetNextRoundTool ? (
+        <StyledRow>
+          <Chip label="請選擇本回合輔助指標" />
+          <StyledOptionsContainer>
+            {Object.values(toolsAvailable).map(tool => (
+              <StyledOptionContainer key={tool.code}>
+                <StyledAccordion
+                  expanded={(expandedTool?.code ?? '') === tool.code}
+                  onChange={() => setExpandedTool(tool)}
+                >
+                  <AccordionSummary>{tool.displayName}</AccordionSummary>
+                  <StyledAccordionDetails>{tool.introduction}</StyledAccordionDetails>
+                </StyledAccordion>
+                {(expandedTool?.code ?? '') === tool.code && (
+                  <IconButton onClick={() => onSelectNextRoundTool(tool)}>
+                    <Done />
+                  </IconButton>
+                )}
+              </StyledOptionContainer>
+            ))}
+          </StyledOptionsContainer>
+        </StyledRow>
+      ) : (
+        <StyledRow>
+          <Chip label="本回合輔助指標" />
+          <div>{selectedTool[selectedTool.length - 1]?.displayName ?? ''}</div>
+          {selectedTool[selectedTool.length - 1]?.introduction && (
+            <HoverTooltip tooltip={selectedTool[selectedTool.length - 1]?.introduction} />
+          )}
+        </StyledRow>
+      )}
       {round > 1 && (
         <>
           <StyledRow>
@@ -33,3 +70,31 @@ export const GameInformationBlock = () => {
     </StyledCard>
   )
 }
+GameInformationBlock.propTypes = {
+  isReadyGetNextRoundTool: PropTypes.bool.isRequired,
+  onSelectNextRoundTool: PropTypes.func.isRequired,
+}
+
+const StyledOptionsContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 0.5rem;
+  width: 100%;
+`
+const StyledOptionContainer = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+`
+const StyledAccordion = styled(Accordion)`
+  background-color: ${props => props.expanded && color.primaryColor400};
+  color: ${props => props.expanded && color.neutralsColor0};
+  width: 17rem;
+`
+const StyledAccordionDetails = styled(AccordionDetails)`
+  display: flex;
+  flex-direction: column;
+  gap: 0.25rem;
+  font-size: 0.75rem;
+`
